Adding parser for 1.4 model manifest
